requirements.txt
pandas
python-telegram-bot>=20.0a0
gspread
gspread-dataframe
gspread-formatting
pyzbar
Pillow
python-dotenv
loguru